refactor: renamed for build circleci
